requirements.txt
duckduckgo-search
rich
python-dotenv
textual
groq
